minor fix on admin delete

+++ main.py
@@ -482,9 +482,10 @@ def delete_combo(combo_id):
     # ratings_to_delete = db.session.query(Rating.rating).filter_by(combo_id=combo_id)
     query = db.session.query(Rating).filter(Rating.combo_id == combo_id)
     ratings_to_delete = query.all()
-    for rating in ratings_to_delete:
-        db.session.delete(rating)
-        db.session.commit()
+    if ratings_to_delete:
+        for rating in ratings_to_delete:
+            db.session.delete(rating)
+            db.session.commit()
     db.session.delete(combo_to_delete)
     db.session.commit()
     return redirect(url_for('view_all'))
